[intf]: add interfaces

+++ clear-hosts.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
+if [[ ! -d /var/run/netns ]]
+then
+    exit 1
+fi
 cd /var/run/netns
-for f in *
+for f in $(ls)
 do
     umount $f
     rm $f
